### 2023-07-07 055bc12
Remove unused references in the BibTeX file.
