+++ joss/paper.bib
@@ -26,27 +26,6 @@
     howpublished = {Retrieved June 30, 2023, from \url{https://gitlab.com/rozsasarpi/rprepo/}}
 }
 
-@article{Tennoe:2018a,
-    author = {S. Tennøe and G. Halnes and G. T. Einevoll},
-    title = {Uncertainpy: A {Python} toolbox for uncertainty quantification and sensitivity analysis in computational neuroscience},
-    journal = {Frontiers in Neuroinformatics},
-    year = 2018,
-    month = aug,
-    volume = 12,
-    number = 49,
-    doi = {10.3389/fninf.2018.00049},
-}
-
-
-@misc{Tennoe:2018b,
-  author = {S. Tennøe},
-  title = {Uncertainpy: A {Python} toolbox for uncertainty quantification and sensitivity analysis in computational neuroscience},
-  year = {2018},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/simetenn/uncertainpy}
-}
-
 @InProceedings{Fekhari2021,
   author    = {Elias Fekhari and Michaël Baudin and Vincent Chabridon and Youssef Jebroun},
   booktitle = {4th International Conference on Uncertainty Quantification in Computational Sciences and Engineering},
