### 2023-07-20 #5671 @editorialbot
[REVIEW]: UQTestFuns: A Python3 library of uncertainty quantification (UQ) test functions
**Submitting author:** @damar-wicaksono (<a href="http://orcid.org/0000-0001-8587-7730">Damar Wicaksono</a>)
**Repository:** https://github.com/damar-wicaksono/uqtestfuns
**Branch with paper.md** (empty if default branch): dev-joss
**Version:** v0.4.1
