### 2023-08-22 damar-wicaksono/uqtestfuns#294 @damar-wicaksono
[JOSS] Add additional usage examples in the docs
Following a feedback in the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/5671#issuecomment-1681454860) additional examples should be added to the repository/docs. I would put these examples in the Getting Started section.
