### 2023-08-22 damar-wicaksono/uqtestfuns#295 @damar-wicaksono
[JOSS] Explicitly states what problems the package aims to solve in the docs
Following a feedback on the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/5671#issuecomment-1681454860), the docs should explicitly state what problems the package aims to solve in the docs. I think such a statement should appear in the landing page of the docs.
