### 2023-08-22 #5671 @damar-wicaksono
@AnjaliSandip, thank you very much for your thorough review and helpful suggestions to improve the package!

I opened two issues in the repo related to your comments above and will deal with them as soon as possible:

- [Add additional usage examples in the docs](https://github.com/damar-wicaksono/uqtestfuns/issues/294)
- [Explicitly states what problems the package aims to solve in the docs](https://github.com/damar-wicaksono/uqtestfuns/issues/295)

And regarding the second author's contribution to the paper and package:

I conceived and coded UQTestFuns with the support of a research project of M. Hecht. Furthermore, M. Hecht's contributions have been indispensable via conceptualization of the package (especially at the early stage), reviewing the high-level interface at the later stage, general supervision of the project (including direction and timeline). Finally, M. Hecht has contributed to the final version of the manuscript. I believe his contributions are sufficient to be a co-author of this submission and hope that this answers your question regarding the contributions of the authors.
