+++ docs/references.bib
@@ -684,4 +684,68 @@ An orthogonal design in which the design matrix has uncorrelated columns is impo
   url         = {http://globalchange.mit.edu/publication/15670},
 }
 
+@Article{Saltelli2002,
+  author  = {Andrea Saltelli},
+  journal = {Computer Physics Communications},
+  title   = {Making best use of model evaluations to compute sensitivity indices},
+  year    = {2002},
+  number  = {2},
+  pages   = {280--297},
+  volume  = {145},
+  doi     = {10.1016/s0010-4655(02)00280-1},
+}
+
+@Article{Li2001,
+  author  = {Genyuan Li and Carey Rosenthal and Herschel Rabitz},
+  journal = {The Journal of Physical Chemistry A},
+  title   = {High dimensional model representations},
+  year    = {2001},
+  number  = {33},
+  pages   = {7765--7777},
+  volume  = {105},
+  doi     = {10.1021/jp010450t},
+}
+
+@Article{Hoeffding1948,
+  author  = {Wassily Hoeffding},
+  journal = {The Annals of Mathematical Statistics},
+  title   = {A class of statistics with asymptotically normal distribution},
+  year    = {1948},
+  number  = {3},
+  pages   = {293--325},
+  volume  = {19},
+  doi     = {10.1214/aoms/1177730196},
+}
+
+@Article{Homma1996,
+  author  = {Toshimitsu Homma and Andrea Saltelli},
+  journal = {Reliability Engineering {\&} System Safety},
+  title   = {Importance measures in global sensitivity analysis of nonlinear models10.1007/978-3-319-11259-6_24-1},
+  year    = {1996},
+  number  = {1},
+  pages   = {1--17},
+  volume  = {52},
+  doi     = {10.1016/0951-8320(96)00002-6},
+}
+
+@Article{Sobol1993,
+  author  = {Sobol', Ilya M.},
+  journal = {Mathematical Modelling and Computational Experiments},
+  title   = {Sensitivity estimates for nonlinear mathematical models},
+  year    = {1993},
+  number  = {4},
+  pages   = {407--414},
+  volume  = {1},
+}
+
+@InCollection{Beck2015,
+  author    = {James L. Beck and Konstantin M. Zuev},
+  booktitle = {Handbook of Uncertainty Quantification},
+  publisher = {Springer International Publishing},
+  title     = {Rare-event simulation},
+  year      = {2015},
+  pages     = {1--26},
+  doi       = {10.1007/978-3-319-11259-6_24-1},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
