+++ docs/getting-started/tutorial-reliability.md
@@ -1,0 +1,316 @@
+---
+jupytext:
+  formats: ipynb,md:myst
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.14.1
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+---
+
+(getting-started:tutorial-reliability)=
+# Tutorial: Reliability Analysis
+
+UQTestFuns includes several test functions from the literature employed
+in a reliability analysis exercise.
+In this tutorial, you'll implement a method to estimate the failure probability
+of a computational model and test the implementation using a test function
+available in UQTestFuns.
+
+By the end of this tutorial, you'll get an idea how a function from UQTestFuns
+is used to test a reliability analysis method.
+
+```{code-cell} ipython3
+import numpy as np
+import matplotlib.pyplot as plt
+import uqtestfuns as uqtf
+```
+
+## Reliability analysis
+
+Consider a system whose performance is defined by a _performance function_[^lsf]
+$g$ whose values, in turn, depend on:
+
+- $\boldsymbol{x}_p$: the (uncertain) input variables
+  of an underlying computational model $\mathcal{M}$
+- $\boldsymbol{x}_s$: additional (uncertain) input variables that affects
+  the performance of the system, but not part of inputs to $\mathcal{M}$ 
+- $\boldsymbol{p}$: an additional set of _deterministic_ parameters of the system
+
+Combining these variables and parameters as an input to the performance function $g$:
+
+$$
+g(\boldsymbol{x}; \boldsymbol{p}) = g(\mathcal{M}(\boldsymbol{x}_p), \boldsymbol{x}_s; \boldsymbol{p}),
+$$
+
+where $\boldsymbol{x} = \{ \boldsymbol{x}_p, \boldsymbol{x}_s \}$.
+
+The system is said to be in _failure state_ if and only if
+$g(\boldsymbol{x}; \boldsymbol{p}) \leq 0$;
+the set of all values $\{ \boldsymbol{x}, \boldsymbol{p} \}$
+such that $g(\boldsymbol{x}; \boldsymbol{p}) \leq 0$
+is called the _failure domain_.
+
+Conversely, the system is said to be in _safe state_
+if and only if $g(\boldsymbol{x}; \boldsymbol{p}) > 0$;
+the set of all values $\{ \boldsymbol{x}, \boldsymbol{p} \}$
+such that $g(\boldsymbol{x}; \boldsymbol{p}) > 0$
+is called the _safe domain_.
+
+### Failure probability
+
+_Reliability analysis_[^rare-event] concerns with estimating
+the failure probability of a system with a given performance function $g$. 
+For a given joint probability density function (PDF) $f_{\boldsymbol{X}}$
+of the uncertain input variables $\boldsymbol{X} = \{ \boldsymbol{X}_p, \boldsymbol{X}_s \}$,
+the failure probability $P_f$ of the system is defined
+as follows {cite}`Sudret2012, Verma2015`:
+
+$$
+P_f \equiv \mathbb{P}[g(\boldsymbol{X}; \boldsymbol{p}) \leq 0] = \int_{\{ \boldsymbol{x} | g(\boldsymbol{x}; \boldsymbol{p}) \leq 0 \}} f_{\boldsymbol{X}} (\boldsymbol{x}) \, \; d\boldsymbol{x}.
+$$
+
+Evaluating the above integral is, in general, non-trivial because the domain
+of integration is only provided implicitly and the number of dimensions may
+be large.
+
+### Monte-Carlo estimation
+
+```{warning}
+The Monte-Carlo method implemented below is one of the most straightforward
[…]
