+++ docs/getting-started/tutorial-sensitivity.md
@@ -1,0 +1,458 @@
+---
+jupytext:
+  formats: ipynb,md:myst
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.14.1
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+---
+
+(getting-started:tutorial-sensitivity)=
+# Tutorial: Sensitivity Analysis
+
+UQTestFuns includes a wide range of test functions from the literature
+employed in a sensitivity analysis exercise.
+In this tutorial, you'll implement a sensitivity analysis method and test
+the implemented method using a test function available in UQTestFuns
+(the popular Ishigami function).
+Afterward, you'll compare the results against analytical results.
+
+By the end of this tutorial, you'll get an idea how a function from UQTestFuns
+is used to test a sensitivity analysis method.
+
+```{code-cell} ipython3
+import numpy as np
+import matplotlib.pyplot as plt
+import uqtestfuns as uqtf
+```
+
+## Global sensitivity analysis
+
+Sensitivity analysis is a model inference technique whose overarching goal is
+to understand the inputs/outputs relationship of a complex
+(potentially, black box) model. Specific goals include:
+
+- _identification of inputs that drive output variability_ that leads to
+  _factor prioritization_, that is, identifying which inputs that result in
+  the largest reduction in the output variability.
+- _identification of non-influential inputs_ that leads to _factor fixing_,
+  that is, identifying which inputs that can be fixed at any value without
+  affecting the output.
+
+A _global_ sensitivity analysis is often distinguished from a _local_ analysis.
+A local analysis is concerned with the effects
+of perturbation around a particular instance of the inputs
+(e.g., their nominal values) on the output.
+A global analysis, on the other hand, is concerned with the effects of
+variations over all possible instances of inputs 
+within their respective uncertainty range on the output.
+
+### Variance decomposition
+
+Consider an $M$-dimensional mathematical function
+$\mathcal{M}: \mathcal{D}_{\boldsymbol{X}} \in [0, 1]^M \mapsto \mathbb{R}$ that represents
+a computational model of interest.
+
+Due to uncertain inputs, represented as an $M$-dimensional random vector $\boldsymbol{X}$,
+the output of the model becomes a random variable $Y = \mathcal{M}(\boldsymbol{X})$.
+It is further assumed that the components $X_m$'s of the random vector $\boldsymbol{X}$
+are mutually independent such that the joint probability density function $f_{\boldsymbol{X}}$
+reads:
+
+$$
+f_{\boldsymbol{X}}(\boldsymbol{x}) = \prod_{m = 1}^M f_{X_m}(x_m),
+$$
+
+where $f_{X_m}$ is the marginal probability density function of $X_m$.
+
+The Hoeffding-Sobol' variance decomposition of random variable $Y$ reads as follows:
+
+$$
+\mathbb{V}[Y] = \sum_{m = 1}^M V_i + \sum_{1 \leq i < j \leq M} V_{m, n} + \ldots + V_{1, \ldots, M},
+$$
+
+where the terms are partial variances defined below:
+
[…]
