### 2023-08-24 e75d568
Reorganize tutorials in the docs.
- Previous getting started guides are now part of the tutorials.
- Add a reference section in "About UQ Test Functions" and
  "Tutorial: Create Built-in Test Functions".
- Fix Issue #296: the keyword `spatial_dimension` is mandatory to
  create a test function with a specified dimension.
