+++ docs/getting-started/tutorials.md
@@ -1,0 +1,11 @@
+(getting-started:tutorials)=
+# UQTestFuns Tutorials
+
+These tutorials are recommended as an introduction to UQTestFuns.
+
+| If you want to...                                                      | Go to...                                                                            |
+|------------------------------------------------------------------------|-------------------------------------------------------------------------------------|
+| get a quick overview of UQTestFuns and its basic usage                 | {ref}`Create Built-in Test Functions <getting-started:tutorial-built-in-functions>` | 
+| learn how to use test functions in the context of sensitivity analysis | {ref}`Test a Sensitivity Analysis Method <getting-started:tutorial-sensitivity>`    | 
+| learn how to use test functions in the context of reliability analysis | {ref}`Test a Reliability Analysis Method <getting-started:tutorial-reliability>`    | 
+| create a custom test function on the fly                               | {ref}`Create a Custom Function <getting-started:tutorial-custom-functions>`         | 
