### 2023-08-24 72a3a1f
Readd JOSS paper to the branch.
