+++ joss/paper.bib
@@ -1,0 +1,257 @@
+@misc{VLSE:2013,
+    author       = {Surjanovic, S. and Bingham, D.},
+    title        = {Virtual Library of Simulation Experiments: Test Functions and Datasets},
+    year         = 2013,
+    howpublished = {Retrieved June 30, 2023, from \url{http://www.sfu.ca/~ssurjano}}
+}
+
+@Misc{GdR:2008,
+  author       = {{GdR MASCOT-NUM}},
+  howpublished = {Retrieved June 30, 2023, from \url{https://www.gdr-mascotnum.fr/benchmarks.html}},
+  title        = {Benchmark proposals of {GdR}},
+  year         = {2008},
+}
+
+@misc{UQWorld:2019,
+    author       = {UQWorld},
+    title        = {Benchmark page of UQWorld, the applied uncertainty quantification community},
+    year         = 2019,
+    howpublished = {Retrieved June 30, 2023, from \url{https://uqworld.org/c/uq-with-uqlab/benchmarks}}
+}
+
+@misc{Rozsas:2019,
+    author       = {Árpád Rózsás and Arthur Slobbe},
+    title        = {Repository and Black-box Reliability Challenge 2019},
+    year         = 2019,
+    howpublished = {Retrieved June 30, 2023, from \url{https://gitlab.com/rozsasarpi/rprepo/}}
+}
+
+@InProceedings{Fekhari2021,
+  author    = {Elias Fekhari and Michaël Baudin and Vincent Chabridon and Youssef Jebroun},
+  booktitle = {4th International Conference on Uncertainty Quantification in Computational Sciences and Engineering},
+  title     = {{otbenchmark}: An open source {Python} package for benchmarking and validating uncertainty quantification algorithms},
+  year      = {2021},
+  doi       = {10.7712/120221.8034.19093},
+  location  = {Athens, Greece},
+}
+
+@Misc{Jakeman2019,
+  author    = {John D. Jakeman},
+  title     = {{PyApprox}},
+  year      = {2019},
+  journal   = {GitHub repository},
+  publisher = {GitHub},
+  url       = {https://github.com/sandialabs/pyapprox},
+}
+
+@Misc{Baudin2021,
+  author    = {Michaël Baudin and Youssef Jebroun and Elias Fekhari and Vincent Chabridon},
+  title     = {{otbenchmark}},
+  year      = {2021},
+  journal   = {GitHub repository},
+  publisher = {GitHub},
+  url       = {https://github.com/mbaudin47/otbenchmark},
+}
+
+@Misc{Bouhlel2023,
+  author    = {Mohamed Amine Bouhlel and John Hwang and Nathalie Bartoli and Rémi Lafage and Joseph Morlier and Joaquim Martins},
+  title     = {{Surrogate Modeling Toolbox}},
+  year      = {2023},
+  journal   = {GitHub repository},
+  publisher = {GitHub},
+  url       = {https://github.com/SMTorg/smt},
+}
+
+@InCollection{Baudin2017,
+  author    = {Michaël Baudin and Anne Dutfoy and Bertrand Iooss and Anne-Laure Popelin},
+  booktitle = {Handbook of Uncertainty Quantification},
+  publisher = {Springer International Publishing},
+  title     = {{OpenTURNS}: An industrial software for uncertainty quantification in simulation},
+  year      = {2017},
+  pages     = {2001--2038},
+  doi       = {10.1007/978-3-319-12385-1_64},
+}
+
+@InProceedings{Sudret2017,
+  author    = {Bruno Sudret and Stefano Marelli and Joe Wiart},
+  booktitle = {2017 11th European Conference on Antennas and Propagation ({EUCAP})},
+  title     = {Surrogate models for uncertainty quantification: An overview},
[…]
