### 2023-08-24 225ee53
Fix Issue #295: Be more explicit about the need for UQTestFuns.
- A statement of need for UQTestFuns is made more explicit in
  the landing page with a link to a more detailed discussion.
- Minor edit on the "About UQ Test Functions" edit.
- Fix Issue #289: remove an unnecessary sentence.
- Fix Issue #288: fix missing subscript in a notation.
