### 2023-08-24 damar-wicaksono/uqtestfuns#295 @damar-wicaksono
This issue has been resolved by PR #300. A paragraph has been added on the landing page of the docs with a link to more detailed discussion regarding the package motivation.
