### 2023-08-24 #5671 @damar-wicaksono
@AnjaliSandip, I tried to address your comments above and updated the documentation of the package. In particular:

- the docs now includes a more explicit statement regarding the need for the package with a link to another part of the docs with more detailed discussion regarding the package motivation (see the [landing page](https://uqtestfuns.readthedocs.io/en/latest/)).
- the docs now includes a set of four tutorials (two were already there but are now reorganized). The new tutorials demonstrate some typical scenario where test functions are used in the context of sensitivity analysis and reliability analysis (see [UQTestFuns Tutorials](https://uqtestfuns.readthedocs.io/en/latest/getting-started/tutorials.html#getting-started-tutorials)).

For comparison the previous docs are still [online](https://uqtestfuns.readthedocs.io/en/stable).

Please let me know of any further comments and feedbacks! Thanks a lot!
