### 2023-08-30 #5671 @Nikoleta-v3
Hey @damar-wicaksono 😄👋🏻 Yup, please go ahead and add the section. Several other publications in JOSS include a contribution section as well.
