### 2023-08-31 damar-wicaksono/uqtestfuns#301 @damar-wicaksono
[JOSS] Add authors' contributions section in the draft manuscript
Following a suggestion on the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/5671#issuecomment-1692523860), a new section regarding the authors' contributions should be added to the draft manuscript.
