### 2023-09-19 damar-wicaksono/uqtestfuns#301 @damar-wicaksono
The section "Acknowledgments" has been updated to "Authors contribution statement and acknowledgments" and now includes the authors' contributions based on the CRediT taxonomy.
