### 2023-10-08 #5671 @Nikoleta-v3
Hey @damar-wicaksono,

I've now reviewed the paper. I can confirm that the package installed smoothly, and I had no issues running the tests. I really appreciate the contributions file. It's very detailed and beginner-friendly.

Although this isn't my field of expertise, from a development standpoint, the package appears to be exceptionally well-developed.

I don't think I've seen another package use Jupyter Books for documentation. I think it works very well! The documentation includes a lot of information, which could be overwhelming, but it's structured very well now. Anja's comment and suggestion regarding this was spot-on.

I'm curious, have you ever used the documentation/package for teaching at your institute or collaborating universities?
