### 2023-10-16 #5671 @damar-wicaksono
Hi @Nikoleta-v3,

Thank you very much for taking the time to review the paper and the package!

> I don't think I've seen another package use Jupyter Books for documentation. I think it works very well!

Thank you! Actually, this is also my first experience of using Jupyter Book to create a documentation for a package (having previously used Sphinx, the community standard). For the current package, I think Jupyter Book provides all what we need out-of-the-box without having to install bunch of supporting plugins.

> The documentation includes a lot of information, which could be overwhelming, but it's structured very well now. Anja's comment and suggestion regarding this was spot-on.

Indeed, we really appreciate @AnjaliSandip's feedbacks on that; it can be difficult for the developer to see which parts are missing in order to make a connection to different parts of the docs, especially if one is new to the package. I'm glad that it now works better, at least, for a knowledgeable new user.

> I'm curious, have you ever used the documentation/package for teaching at your institute or collaborating universities?

For teaching per se, not yet. But one of the reasons we are developing the package is also because when we introduce a method for, say, sensitivity analysis in a classroom setting we can conveniently use one of the known test functions from the literature implemented in the package as an example or illustration.

That being said, there's a student project in our group (an investigation of a metamodeling/surrogate modeling method) that employed the test functions of the package to benchmark the proposed method. Convenient access to test functions for testing is one of the main goals of the package.
