### 2023-10-17 #5671 @Nikoleta-v3
ChatGPT
Hello @damar-wicaksono. Thank you very much for your reply!

> For teaching per se, not yet. But one of the reasons we are developing the package is also because when we introduce a method for, say, sensitivity analysis in a classroom setting, we can conveniently use one of the known test functions from the literature implemented in the package as an example or illustration. We are planning to have such a session in our institute in the future.

That's amazing. I think it would be great to eventually have a dedicated section in the documentation about the various ways you can use the package in teaching, based on your experiences.

> That being said, there's a student project in our group (an investigation of a metamodeling/surrogate modeling method) that employed the test functions of the package to benchmark the proposed method. Convenient access to test functions for testing is one of the main goals of the package.

And, as you mentioned here, eventually see the package being cited in such projects!

I am very happy to say that I have no further comments at this point. I have already spoken to your second reviewer, and they should be reviewing the paper soon. Thank you again for your patience! 🙏🏻
