### 2023-10-20 #5671 @Himscipy
Hello @damar-wicaksono,
I found UQTestFuns an interesting software tool for training students on doing UQ for scientific applications.  Based on the current implementation it will require additional capabilities to be generalized for doing UQ for black-box functions or complex scientific application.

The package is well documented and checks all the review criteria according to my review.  I agree with @Nikoleta-v3  that the documentation is well structured to guide the new user about the package. In addition, the revisions made based on  @AnjaliSandip comments have improved the paper and the repository.

I have completed my review and accept the repo for publication.  Thank you again for your patience !
