### 2023-10-23 #5671 @Nikoleta-v3
@damar-wicaksono  you please:

- [ ] (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release 0.4.0
 Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
