### 2023-10-27 9b0f99b
Add an implementation of the low-dim high-deg test function
The two-dimensional polynomial test function of high-degree from
Alemazkoor and Meidani (2008) has been implemented.
