+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = uqtestfuns
-version = 0.4.0
+version = 0.4.1
 url = https://github.com/damar-wicaksono/uqtestfuns
 author = Damar Wicaksono
 author_email = damar.wicaksono@outlook.com
