### 2023-10-30 #5671 @Nikoleta-v3
Hey @damar-wicaksono one of the references is causing a warning: https://github.com/openjournals/joss-reviews/issues/5671#issuecomment-1784607971.

The journal should be "Reliability Engineering & System Safety" could you tweak [this line](https://github.com/damar-wicaksono/uqtestfuns/blob/790c7ebe181edabd07299aba410adbdac9211bad/joss/paper.bib#L227C1-L227C1).
