+++ joss/paper.bib
@@ -224,7 +224,7 @@
 
 @Article{Castellon2023,
   author  = {Dario Fernandez Castellon and Aksel Fenerci and {\O}yvind Wiig Petersen and Ole {\O}iseth},
-  journal = {Reliability Engineering {\&}amp$\mathsemicolon$ System Safety},
+  journal = {Reliability Engineering \& System Safety},
   title   = {Full long-term buffeting analysis of suspension bridges using {Gaussian} process surrogate modelling and importance sampling {Monte Carlo} simulations},
   year    = {2023},
   volume  = {235},
