### 2023-10-30 210679b
minor changes in paper
