### 2023-10-30 #5671 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
