### 2023-10-30 #5671 @danielskatz
Congratulations to @damar-wicaksono (Damar Wicaksono) and co-author on your publication!!

And thanks to @Himscipy and @AnjaliSandip for reviewing, and to @Nikoleta-v3
JOSS is dependent on volunteers and we couldn't do this without you
